==> notebooks/01_descriptors_computation.ipynb <==
# Descriptors Computation
This notebook handles the descriptors computation. 
This procedure involves the following steps:
- For each variable, we estimate its Markov Blanket (MB) by selecting its lagged versions from one time step before and one after. 
- We standardize the time series to avoid varsortability
- Using the estimated MB, we compute a set of descriptors for all possible causal pairs (i.e., $ t-\tau \rightarrow t, \forall \tau$). that characterize the causal relationship between the variable pairs. These descriptors include conditional mutual information terms and other statistical properties that provide insights into the dependencies and interactions between the variables.
- For families of descriptors, we compute the quantiles of their empirical distributions. This step captures the distributional characteristics and aids in feature representation for the classifier.
- The computed descriptors and their quantiles are compiled into an input feature vector. This vector encapsulates the essential characteristics of the causal relationships and serves as the input for the classifier.
- For training data, each input vector is labeled as causal (1) or noncausal (0) based on the original selection criteria from the synthetic data's Directed Acyclic Graph (DAG). This labeling is crucial for supervised learning and model training.
- The labeled dataset, comprising the feature vectors, is used to train a classifier. The classifier learns to predict the likelihood of causal relationships based on the descriptors.
- For unseen time series data, the trained classifier predicts the probability of causal links for each pair of variables. The predictions are based on the computed descriptors for the test data.

Here are the descriptors 
\begin{align}
n  \\
m \\
m/n \\
b: \mathbf{z}_j = b \cdot (\mathbf{z}_i \oplus \mathbf{MB}_j) \text{, where} \oplus\text{indicates vector concatenation}  \\
b: \mathbf{z}_i = b \cdot (\mathbf{z}_j \oplus \mathbf{MB}_i) \text{, where} \oplus\text{indicates vector concatenation}   \\
\operatorname{kurt}(\mathbf{z}_i) = \mathbb{E}[(\mathbf{z}_i-\mathbb{E}[\mathbf{z}_i])^4] / \left(\mathbb{E}[(\mathbf{z}_i-\mathbb{E}[\mathbf{z}_i])^2]\right)^2 - 3 \\ 
\operatorname{kurt}(\mathbf{z}_j) = \mathbb{E}[(\mathbf{z}_j-\mathbb{E}[\mathbf{z}_j])^4] / \left(\mathbb{E}[(\mathbf{z}_j-\mathbb{E}[\mathbf{z}_j])^2]\right)^2 - 3 \\ 
\operatorname{skew}(\mathbf{z}_i) = \mathbb{E}[(\mathbf{z}_i-\mathbb{E}[\mathbf{z}_i])^3] / \left(\mathbb{E}[(\mathbf{z}_i-\mathbb{E}[\mathbf{z}_i])^2]\right)^{3/2} \\
\operatorname{skew}(\mathbf{z}_j) = \mathbb{E}[(\mathbf{z}_j-\mathbb{E}[\mathbf{z}_j])^3] / \left(\mathbb{E}[(\mathbf{z}_j-\mathbb{E}[\mathbf{z}_j])^2]\right)^{3/2}
 \\
\operatorname{HOC_{1,2}}(\mathbf{z}_i, \mathbf{z}_j) = \mathbb{E}\left[ (\mathbf{z}_i - \mathbb{E}[\mathbf{z}_i])^1 \cdot (\mathbf{z}_j - \mathbb{E}[\mathbf{z}_j])^2 \right] \\
\operatorname{HOC_{2,1}}(\mathbf{z}_i, \mathbf{z}_j) = \mathbb{E}\left[ (\mathbf{z}_i - \mathbb{E}[\mathbf{z}_i])^2 \cdot (\mathbf{z}_j - \mathbb{E}[\mathbf{z}_j])^1 \right] \\
\operatorname{HOC_{1,3}}(\mathbf{z}_i, \mathbf{z}_j) = \mathbb{E}\left[ (\mathbf{z}_i - \mathbb{E}[\mathbf{z}_i])^1 \cdot (\mathbf{z}_j - \mathbb{E}[\mathbf{z}_j])^3 \right] \\
\operatorname{HOC_{3,1}}(\mathbf{z}_i, \mathbf{z}_j) = \mathbb{E}\left[ (\mathbf{z}_i - \mathbb{E}[\mathbf{z}_i])^3 \cdot (\mathbf{z}_j - \mathbb{E}[\mathbf{z}_j])^1 \right] \\
    I(\mathbf{z}_i ; \mathbf{z}_j)  \\
    I(\mathbf{z}_j ; \mathbf{z}_i)  \\ 
    I(\mathbf{m}_j^{(k)} ; \mathbf{z}_i) \forall \mathbf{m}_j^{(k)} \in \mathbf{MB}_j  \\
    I(\mathbf{m}_i^{(k)} ; \mathbf{z}_j) \forall \mathbf{m}_i^{(k)} \in \mathbf{MB}_i \\
    I(\mathbf{z}_i ; \mathbf{z}_j | \mathbf{MB}_i \cap \mathbf{MB}_j)  \\
    I(\mathbf{z}_j ; \mathbf{z}_i | \mathbf{MB}_j)  \\
    I(\mathbf{z}_i ; \mathbf{z}_j | \mathbf{MB}_i)  \\
    I(\mathbf{z}_j ; \mathbf{z}_i | \mathbf{MB}_i \cup \mathbf{m}_j^{(k)}) \forall \mathbf{m}_j^{(k)} \in \mathbf{MB}_j  \\
    I(\mathbf{z}_i ; \mathbf{z}_j | \mathbf{MB}_j \cup \mathbf{m}_i^{(k)}) \forall \mathbf{m}_i^{(k)} \in \mathbf{MB}_i \\
    I(\mathbf{z}_i ; \mathbf{m}_j^{(k)} | \mathbf{z}_j) \forall \mathbf{m}_j^{(k)} \in \mathbf{MB}_j  \\
I(\mathbf{z}_j ; \mathbf{m}_i^{(k)} | \mathbf{z}_i) \forall \mathbf{m}_i^{(k)} \in \mathbf{MB}_i  \\
    I(\mathbf{m}_i^{(k_i)} ; \mathbf{m}_j^{(k_j)} | \mathbf{z}_i) \forall (\mathbf{m}_i^{(k_i)}, \mathbf{m}_j^{(k_j)}) \in \mathbf{MB_{i}} \times \mathbf{MB_{j}}  \\
    I(\mathbf{m}_j^{(k_j)} ; \mathbf{m}_i^{(k_i)} | \mathbf{z}_j) \forall (\mathbf{m}_i^{(k_i)}, \mathbf{m}_j^{(k_j)}) \in \mathbf{MB_{i}} \times \mathbf{MB_{j}} \\
        I(\mathbf{m}_i^{(k_{i}^{1})} ; \mathbf{m}_i^{(k_{i}^{2})} | \mathbf{z}_i) - I(\mathbf{m}_i^{(k_{i}^{1})} ; \mathbf{m}_i^{(k_{i}^{2})}) \forall (\mathbf{m}_i^{(k_{i}^{1})}, \mathbf{m}_i^{(k_{i}^{2})}) \in \mathbf{MB_{i}} \times \mathbf{MB_{i}}  \\
    I(\mathbf{m}_j^{(k_{j}^{1})} ; \mathbf{m}_j^{(k_{j}^{2})} | \mathbf{z}_j) - I(\mathbf{m}_j^{(k_{j}^{1})} ; \mathbf{m}_j^{(k_{j}^{2})}) \forall (\mathbf{m}_j^{(k_{j}^{1})}, \mathbf{m}_j^{(k_{j}^{2})}) \in \mathbf{MB_{j}} \times \mathbf{MB_{j}}     
\end{align}

from d2c.descriptors import D2C, DataLoader

N_VARS = 5
MAXLAGS = 3
N_JOBS = 40

The DataLoader class for handling time series data and directed acyclic graphs (DAGs) to prepare for the following stage: the descriptors computation.
The preparation includes:
1. Creating lagged time series: This step involves generating lagged versions of the original time series data. Lagged time series help in capturing temporal dependencies and interactions between variables at different time steps.
2. Flattening the original dictionaries into coherent lists: The original data, which may be stored in nested dictionaries, is flattened into lists. This transformation ensures that the data is in a consistent and accessible format for further processing.
3. Renaming the nodes of the DAGs: The nodes of the Directed Acyclic Graphs (DAGs) are renamed to maintain consistency and clarity. This step is crucial for accurately representing the causal relationships between variables in the DAGs.


dataloader = DataLoader(n_variables = N_VARS,
                    maxlags = MAXLAGS)
dataloader.from_pickle('example/synthetic_data_train.pkl')

original_observations = dataloader.get_original_observations()
lagged_flattened_observations = dataloader.get_observations()
flattened_dags = dataloader.get_dags()

We are now ready for the core of our methodology: the D2C method. <br>
This method starts from a list of observations and dags and computes the corresponding descritpors, storing them in a dataframe. <br>
The D2C class gets the following arguments: 
- `observations` (list): List of observations (pd.DataFrame) corresponding to each DAG.
- `dags` (list): List of directed acyclic graphs (DAGs) representing causal relationships.
- `couples_to_consider_per_dag` (int, optional): To speedup, one can consider only a limited number of possible couples of variables. Couples are chosen to respect a specific ratio of causal/noncausal. Therefore, a DAG must be available: it can only be used for training data. For testing purposes only, we recommend using `D2CWrapper` instead. Defaults to -1 (all couples).  
- `n_variables` (int, optional): Number of variables in the time series. Defaults to 3.
- `maxlags` (int, optional): Maximum number of lags in the time series. Defaults to 3.
- `seed` (int, optional): Random seed for reproducibility. Defaults to 42.
- `n_jobs` (int, optional): Number of parallel jobs to run. Defaults to 1.
 - `full` (bool, optional): computes the whole set of descriptors rather than just a subset. Defaults to True. 



d2c = D2C(observations=lagged_flattened_observations,
        dags=flattened_dags, 
        couples_to_consider_per_dag=20, 
        n_variables=N_VARS, 
        maxlags=MAXLAGS,
        seed=42,
        n_jobs=N_JOBS,
        full=True)

d2c.initialize()

Now let's look at the descriptors. 

We remind our variable naming convention: <br>
- A time series of `n_variables` dimensions will have names from 0 to `n_variables - 1` to refer to the variables at time `t` (present)
- names from `n_variables` to `n_variables*2 - 1` will indicate the same variable at time `t-1` (1-lag)
- names from `n_variables*2` to `n_variables*3 - 1` will indicate the same variable at time `t-2` (2-lag)  <br>
For example if `n_variables = 5`, the line where `edge_source` is 12 and `edge_dest` is 4, refers to the link between variable `3` at `t-2` to variable `5` at time `t`

descriptors_df_train = d2c.get_descriptors_df()
print(descriptors_df_train.columns)
descriptors_df_train

Now we can see if we can learn something from these descriptors. We will perform a 50% train/test split and run a classifier. 

dataloader = DataLoader(n_variables = N_VARS,
                    maxlags = MAXLAGS)
dataloader.from_pickle('example/synthetic_data_test.pkl')

d2c = D2C(observations=dataloader.get_observations(),
        dags=dataloader.get_dags(), 
        couples_to_consider_per_dag=20, 
        n_variables=N_VARS, 
        maxlags=MAXLAGS,
        seed=42,
        n_jobs=N_JOBS,
        full=True)

d2c.initialize()

descriptors_df_test = d2c.get_descriptors_df()

X_train = descriptors_df_train.drop(columns=['graph_id','edge_source','edge_dest','is_causal'])
y_train = descriptors_df_train['is_causal']
X_test = descriptors_df_test.drop(columns=['graph_id','edge_source','edge_dest','is_causal'])
y_test = descriptors_df_test['is_causal']

from imblearn.ensemble import BalancedRandomForestClassifier
from sklearn.metrics import accuracy_score, roc_auc_score, confusion_matrix

clf = BalancedRandomForestClassifier(n_estimators=10, max_depth=None, random_state=0, sampling_strategy='auto',replacement=True,bootstrap=True)
clf.fit(X_train, y_train)
y_pred = clf.predict(X_test)
y_pred_proba = clf.predict_proba(X_test)

print(f"Accuracy: {accuracy_score(y_test, y_pred):.4f}")
print(f"ROC AUC Score: {roc_auc_score(y_test, y_pred_proba[:, 1]):.4f}")
print("Confusion Matrix:")
print(confusion_matrix(y_test, y_pred))

Scores are promising. We now save the descriptors.

# save the descriptors_df to disk
import pickle
filename = 'example/descriptors_df_train.pkl'
pickle.dump(descriptors_df_train, open(filename, 'wb'))
